+++ package.json
@@ -65,7 +65,6 @@
     "redux-first-router": "^2.1.1",
     "redux-first-router-link": "^2.1.1",
     "redux-first-router-restore-scroll": "^1.2.2",
-    "redux-saga": "^1.0.0",
     "redux-thunk": "^2.3.0",
     "request": "^2.88.0",
     "reselect": "^4.0.0",
