+++ packages/earth-map/src/modules/global/actions.ts
@@ -19,5 +19,4 @@
 
 import { createAction } from 'vizzuality-redux-tools';
 
-export const persistData = createAction('GLOBAL/persistData');
 export const setLastViewedPlace = createAction('GLOBAL/setLastViewedPlace');
