+++ packages/earth-map/src/store/index.tsx
@@ -20,7 +20,6 @@
 import { applyMiddleware, combineReducers, createStore, Store } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import { connectRoutes } from 'redux-first-router';
-import createSagaMiddleware from 'redux-saga';
 import thunk from 'redux-thunk';
 import { handleModule } from 'vizzuality-redux-tools';
 
@@ -37,10 +36,7 @@ import * as user from '../modules/user';
 import * as widget from '../modules/widget';
 import * as widgets from '../modules/widgets';
 import { CONFIG, ROUTES } from '../routes';
-import sagas from '../sagas';
-import restoreState from '../store/ephemeral-state';
-
-const sagaMiddleware = createSagaMiddleware();
+import initEphemeralState from '../store/ephemeral-state';
 
 const initStore = (initialState = {}) => {
   // Create router reducer, middleware and enhancer
@@ -84,7 +80,7 @@ const initStore = (initialState = {}) => {
     traceLimit: 10,
   });
 
-  const middlewares = applyMiddleware(thunk, routerMiddleware, sagaMiddleware);
+  const middlewares = applyMiddleware(thunk, routerMiddleware);
   const enhancers = composeEnhancer(routerEnhancer, middlewares);
 
   // create store
@@ -93,10 +89,8 @@ const initStore = (initialState = {}) => {
 
   // restore state from sessionStorage
   const ephemeralState = SessionStorage.getObject('ephemeral');
-  restoreState(store, ephemeralState);
+  initEphemeralState(store, ephemeralState);
 
-  // run the sagas && initialDispatch
-  sagaMiddleware.run(sagas);
   initialDispatch();
 
   return { store };
